push

+++ ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: abchikhi <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/02 17:29:09 by abchikhi          #+#    #+#             */
-/*   Updated: 2023/11/02 17:29:11 by abchikhi         ###   ########.fr       */
+/*   Updated: 2023/11/15 00:29:49 by abchikhi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,8 +19,8 @@ static int	is_space(char c)
 
 int	ft_atoi(const char *str)
 {
-	int	res;
-	int	sign;
+	long	res;
+	int		sign;
 
 	res = 0;
 	sign = 1;
@@ -39,3 +39,12 @@ int	ft_atoi(const char *str)
 	}
 	return (res * sign);
 }
+
+// #include <string.h>
+// #include <stdio.h>
+
+// int main()
+// {
+// 	// printf("%d\n", ft_atoi(NULL));
+// 	printf("%d\n", atoi(NULL));
+// }
